Move ResultWithTotal object from services to utils.
