+++ internal/utils/result_with_total.go
@@ -1,4 +1,4 @@
-package services
+package utils
 
 type ResultWithTotal struct {
 	Total  int
