+++ internal/utils/result_with_total_test.go
@@ -1,4 +1,4 @@
-package services
+package utils
 
 import (
 	"testing"
